Refactoring

+++ restaurant-booking-api/src/main/java/com/alaasmagi/restaurant_booking_api/application/contracts/ITableRepository.java
@@ -5,5 +5,5 @@ import com.alaasmagi.restaurant_booking_api.domain.TableEntity;
 import java.util.UUID;
 
 public interface ITableRepository extends IRepository<TableEntity, UUID> {
-    TableEntity changePosition(UUID id, double x, double y);
+    TableEntity changePosition(UUID id, int x, int y);
 }
